app/requirements.txt
africastalking==1.2.0
alembic==1.4.1
amqp==2.5.2
astroid==2.3.3
attrs==19.3.0
bcrypt==3.1.7
billiard==3.6.3.0
blinker==1.4
celery==4.4.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
contextlib2==0.5.5
cryptography==2.8
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
idna==2.9
importlib-metadata==1.5.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
kombu==4.6.8
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
phonenumbers==8.11.5
psycopg2-binary==2.8.4
pycparser==2.20
PyJWT==1.7.1
pylint==2.4.4
pyotp==2.3.0
pyrsistent==0.16.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
redis==3.4.1
requests==2.23.0
schema==0.7.1
six==1.14.0
SQLAlchemy==1.3.15
toastedmarshmallow==2.15.2.post1
typed-ast==1.4.1
urllib3==1.25.8
vine==1.3.0
Werkzeug==1.0.0
wrapt==1.11.2
zipp==3.1.0
pre-commit==2.4.0
